requirements.txt
cloudinary==1.3.2
dj-database-url==0.4.0
Django==1.9.6
djangorestframework==3.3.3
gunicorn==19.4.5
Pillow==3.2.0
psycopg2==2.6.1
SQLAlchemy==1.0.13
virtualenv==15.0.2
whitenoise==2.0.6
